requirements.txt
numpy==1.24.3
matplotlib==3.1.1
tqdm==4.65.0
torchvision==0.13.0
torch==1.12.0
Pillow==9.5.0
sklearn
clip